requirements.txt
aiogram==2.12.1
aiohttp==3.7.4.post0
astroid==2.5.2
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
environs==9.3.2
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
marshmallow==3.11.1
mccabe==0.6.1
multidict==5.1.0
oauthlib==3.1.0
packaging==20.9
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.7.4
pyparsing==2.4.7
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
wrapt==1.12.1
yarl==1.6.3
